# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "CallBody": {
        "description": "``POST /call`` 的请求体。\n\n必须定义在**模块级**：本模块用了 ``from __future__ import annotations``，\n若把它定义在 ``create_app`` 内部，FastAPI 解析注解时会去模块全局找而找不到，\n于是把 ``body`` 误判成查询参数。",
        "properties": {
          "args": {
            "additionalProperties": true,
            "description": "工具参数",
            "title": "Args",
            "type": "object"
          },
          "name": {
            "description": "工具名，例如 ppt_open",
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CallBody",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PaceBody": {
        "description": "``POST /pace`` 的请求体。同样必须定义在模块级，理由见 ``CallBody``。",
        "properties": {
          "pace_ms": {
            "description": "写操作之间的停顿毫秒数；0=极速，400≈正常，1500=慢速",
            "maximum": 10000.0,
            "minimum": 0.0,
            "title": "Pace Ms",
            "type": "integer"
          }
        },
        "required": [
          "pace_ms"
        ],
        "title": "PaceBody",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "ppt_open_body": {
        "additionalProperties": false,
        "properties": {
          "bring_to_front": {
            "default": true,
            "description": "打开后把 PowerPoint 窗口切到最前，让用户看得见",
            "title": "Bring To Front",
            "type": "boolean"
          },
          "path": {
            "description": "演示文件路径（.pptx / .ppt）",
            "minLength": 1,
            "title": "Path",
            "type": "string"
          },
          "read_only": {
            "default": false,
            "description": "以只读方式打开，避免误改原文件",
            "title": "Read Only",
            "type": "boolean"
          }
        },
        "required": [
          "path"
        ],
        "title": "ppt_open_body",
        "type": "object"
      },
...
      "ppt_shot_body": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "default": 1,
            "description": "连续导出几页",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Count",
            "type": "integer"
          },
          "deck": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "演示名、路径或序号；留空用当前活动演示",
            "title": "Deck"
          },
          "format": {
            "default": "png",
            "description": "高清图格式：png 清晰、jpg 更小",
            "enum": [
              "png",
              "jpg"
            ],
            "title": "Format",
            "type": "string"
          },
          "goto": {
            "default": true,
            "description": "截图后把视图切到第一张，便于用户跟随",
            "title": "Goto",
            "type": "boolean"
          },
          "inline": {
            "default": true,
            "description": "是否生成给模型内联查看的缩略图",
            "title": "Inline",
            "type": "boolean"
          },
          "slide": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "起始页码；留空用当前正在看的那页",
            "title": "Slide"
          },
          "thumb_width": {
            "default": 1024,
            "description": "给模型看的缩略图宽度（像素）",
            "maximum": 1600.0,
            "minimum": 320.0,
            "title": "Thumb Width",
            "type": "integer"
          },
          "width": {
            "default": 1280,
            "description": "高清图宽度（像素）",
            "maximum": 3840.0,
            "minimum": 320.0,
            "title": "Width",
            "type": "integer"
          }
        },
        "title": "ppt_shot_body",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "让 agent 实时控制正在运行的 PowerPoint。\n\n同一套工具同时暴露为 MCP 与 HTTP/OpenAPI；所有操作都会广播到观察台，用户随时看得见 agent 在做什么。\n\n**鉴权**：除 `/health` 外都需要令牌，用 `X-PPT-Token` 头或 `?token=` 传入；令牌在守护进程的 runtime.json 里，用 `pptctl url` 可打印观察台地址。",
    "title": "ppt-agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/call": {
      "post": {
        "operationId": "call_call_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "按名字调用任意工具",
        "tags": [
          "invoke"
        ]
      }
    },
...
    "/feed": {
      "get": {
        "description": "一次请求拿全面板要显示的东西。\n\nSSE（``/events``）适合网页；但 VBA、PowerShell 这类客户端不方便消费长连接，\n所以同样一份数据再给一个轮询端点——PowerPoint 里的加载项面板就靠它。",
        "operationId": "feed_feed_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 40,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Feed Feed Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "面板快照：状态 + 最近的操作流水",
        "tags": [
          "meta"
        ]
      }
    },
    "/feed.txt": {
      "get": {
        "description": "与 ``/feed`` 同一份数据，但用**纯文本**。\n\n为什么要有这个：VBA 里解析 JSON 要手写解析器，脆弱又长。纯文本格式\n一行一条、用 `` | `` 分隔，VBA 一句 ``Split`` 就能用；PowerShell 里\n``curl`` 直接看得懂，调试也方便。",
        "operationId": "feed_text_feed_txt_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 40,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "面板快照（纯文本，给 VBA/PowerShell 用）",
        "tags": [
          "meta"
        ]
      }
    },
...
    "/pace": {
      "post": {
        "operationId": "set_pace_pace_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaceBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Set Pace Pace Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "设置跟速（观察台与任务窗格用）",
        "tags": [
          "meta"
        ]
      }
    },
    "/shots/{name}": {
      "get": {
        "operationId": "shot_file_shots__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "取一张截图文件（观察台与任务窗格用）",
        "tags": [
          "meta"
        ]
      }
    },
...
    "/tools/ppt_open": {
      "post": {
        "description": "打开一个演示文稿并让它在 PowerPoint 里显示出来\n\n- 行为：`write`（read=只读，write=写入，destroy=破坏性）\n- 返回：被打开演示的摘要（页数、路径、是否只读）",
        "operationId": "ppt_open",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ppt_open_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "成功"
          },
          "400": {
            "description": "工具报错（响应体里带 error.code / error.message）"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "打开一个演示文稿并让它在 PowerPoint 里显示出来",
        "tags": [
          "write"
        ]
      }
    },
...
    "/tools/ppt_shot": {
      "post": {
        "description": "把某一页导出成图片：给人看高清 PNG，给模型看压缩缩略图\n\n- 行为：`read`（read=只读，write=写入，destroy=破坏性）\n- 返回：每页的高清图片与缩略图（含路径与可访问 URL）",
        "operationId": "ppt_shot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ppt_shot_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "成功"
          },
          "400": {
            "description": "工具报错（响应体里带 error.code / error.message）"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "把某一页导出成图片：给人看高清 PNG，给模型看压缩缩略图",
        "tags": [
          "read"
        ]
      }
    },
...
  }
}